Drop support for Postgres 9
