+++ psalm.xml.dist
@@ -46,11 +46,6 @@
                     is no longer supported.
                 -->
                 <file name="src/Tools/Console/ConsoleRunner.php"/>
-                <!--
-                    TODO: remove in 4.0.0
-                -->
-                <referencedClass name="Doctrine\DBAL\Platforms\Keywords\PostgreSQL100Keywords"/>
-                <referencedClass name="Doctrine\DBAL\Platforms\PostgreSQL100Platform"/>
             </errorLevel>
         </DeprecatedClass>
         <DeprecatedMethod>
