--- src/Platforms/Keywords/PostgreSQL100Keywords.php
@@ -1,18 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace Doctrine\DBAL\Platforms\Keywords;
-
-/**
- * PostgreSQL 10.0 reserved keywords list.
- *
- * @deprecated Use {@link PostgreSQLKeywords} instead.
- */
-class PostgreSQL100Keywords extends PostgreSQLKeywords
-{
-    public function getName(): string
-    {
-        return 'PostgreSQL100';
-    }
-}
