--- src/Platforms/PostgreSQL100Platform.php
@@ -1,34 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace Doctrine\DBAL\Platforms;
-
-use Doctrine\DBAL\Platforms\Keywords\KeywordList;
-use Doctrine\DBAL\Platforms\Keywords\PostgreSQL100Keywords;
-
-/**
- * Provides the behavior, features and SQL dialect of the PostgreSQL 10.0 database platform.
- *
- * @deprecated This class will be merged with {@see PostgreSQLPlatform} in 4.0 because support for Postgres
- *             releases prior to 10.0 will be dropped.
- */
-class PostgreSQL100Platform extends PostgreSQLPlatform
-{
-    protected function createReservedKeywordsList(): KeywordList
-    {
-        return new PostgreSQL100Keywords();
-    }
-
-    public function getListSequencesSQL(string $database): string
-    {
-        return 'SELECT sequence_name AS relname,
-                       sequence_schema AS schemaname,
-                       minimum_value AS min_value,
-                       increment AS increment_by
-                FROM   information_schema.sequences
-                WHERE  sequence_catalog = ' . $this->quoteStringLiteral($database) . "
-                AND    sequence_schema NOT LIKE 'pg\_%'
-                AND    sequence_schema != 'information_schema'";
-    }
-}
