+++ src/Platforms/PostgreSQLPlatform.php
@@ -167,10 +167,13 @@ class PostgreSQLPlatform extends AbstractPlatform
 
     public function getListSequencesSQL(string $database): string
     {
-        return "SELECT sequence_name AS relname,
-                       sequence_schema AS schemaname
+        return 'SELECT sequence_name AS relname,
+                       sequence_schema AS schemaname,
+                       minimum_value AS min_value,
+                       increment AS increment_by
                 FROM   information_schema.sequences
-                WHERE  sequence_schema NOT LIKE 'pg\_%'
+                WHERE  sequence_catalog = ' . $this->quoteStringLiteral($database) . "
+                AND    sequence_schema NOT LIKE 'pg\_%'
                 AND    sequence_schema != 'information_schema'";
     }
