+++ tests/Driver/AbstractPostgreSQLDriverTest.php
@@ -10,7 +10,6 @@ use Doctrine\DBAL\Driver\AbstractPostgreSQLDriver;
 use Doctrine\DBAL\Driver\API\ExceptionConverter;
 use Doctrine\DBAL\Driver\API\PostgreSQL;
 use Doctrine\DBAL\Platforms\AbstractPlatform;
-use Doctrine\DBAL\Platforms\PostgreSQL100Platform;
 use Doctrine\DBAL\Platforms\PostgreSQLPlatform;
 use Doctrine\DBAL\Schema\AbstractSchemaManager;
 use Doctrine\DBAL\Schema\PostgreSQLSchemaManager;
@@ -20,6 +19,11 @@ use Doctrine\DBAL\Schema\PostgreSQLSchemaManager;
  */
 class AbstractPostgreSQLDriverTest extends AbstractDriverTest
 {
+    public function testThrowsExceptionOnCreatingDatabasePlatformsForInvalidVersion(): void
+    {
+        self::markTestSkipped('PostgreSQL drivers do not use server version to instantiate platform');
+    }
+
     protected function createDriver(): Driver
     {
         return $this->getMockForAbstractClass(AbstractPostgreSQLDriver::class);
@@ -49,10 +53,9 @@ class AbstractPostgreSQLDriverTest extends AbstractDriverTest
     public static function platformVersionProvider(): array
     {
         return [
-            ['9.4', PostgreSQLPlatform::class],
-            ['9.4.0', PostgreSQLPlatform::class],
-            ['9.4.1', PostgreSQLPlatform::class],
-            ['10', PostgreSQL100Platform::class],
+            ['10.0', PostgreSQLPlatform::class],
+            ['11.0', PostgreSQLPlatform::class],
+            ['13.3', PostgreSQLPlatform::class],
         ];
     }
 }
