+++ tests/Platforms/PostgreSQLPlatformTest.php
@@ -1027,4 +1027,19 @@ class PostgreSQLPlatformTest extends AbstractPlatformTestCase
         self::assertTrue($this->platform->hasDoctrineTypeMappingFor('jsonb'));
         self::assertEquals(Types::JSON, $this->platform->getDoctrineTypeMapping('jsonb'));
     }
+
+    public function testGetListSequencesSQL(): void
+    {
+        self::assertSame(
+            "SELECT sequence_name AS relname,
+                       sequence_schema AS schemaname,
+                       minimum_value AS min_value,
+                       increment AS increment_by
+                FROM   information_schema.sequences
+                WHERE  sequence_catalog = 'test_db'
+                AND    sequence_schema NOT LIKE 'pg\_%'
+                AND    sequence_schema != 'information_schema'",
+            $this->platform->getListSequencesSQL('test_db')
+        );
+    }
 }
